Inventory Management Implemented . Product Refactord

+++ InventoryManagement.Infrastructure.EFCore/Migrations/InventoryContextModelSnapshot.cs
@@ -39,8 +39,8 @@ namespace InventoryManagement.Infrastructure.EFCore.Migrations
                     b.Property<long>("ProductId")
                         .HasColumnType("bigint");
 
-                    b.Property<long>("UnitPrice")
-                        .HasColumnType("bigint");
+                    b.Property<double>("UnitPrice")
+                        .HasColumnType("float");
 
                     b.HasKey("Id");
